Refactor and add tests
