+++ src/scenes/Moves/Ppm/api.js
@@ -1,4 +1,4 @@
-import { getClient, checkResponse, getPublicClient } from 'shared/Swagger/api';
+import { getClient, checkResponse } from 'shared/Swagger/api';
 import { formatPayload, formatDateString } from 'shared/utils';
 
 export async function GetPpm(moveId) {
@@ -59,16 +59,6 @@ export async function GetPpmSitEstimate(moveDate, sitDays, originZip, destZip, w
     weight_estimate: weightEstimate,
   });
   checkResponse(response, 'failed to update ppm due to server error');
-  return response.body;
-}
-
-export async function ValidateZipRateData(zipCode, zipType) {
-  const client = await getPublicClient();
-  const response = await client.apis.postal_codes.validatePostalCode({
-    postal_code: zipCode,
-    postal_code_type: zipType,
-  });
-  checkResponse(response, 'failed to validate ppm data due to server error');
   return response.body;
 }
